stop using em-synchrony as base adapter
