+++ Appraisals
@@ -2,4 +2,4 @@
 
 require "appraisal/matrix"
 
-appraisal_matrix(rails: [">= 6.1", "< 7.1"])
+appraisal_matrix(rails: [">= 7.0", "< 7.1"])
