--- gemfiles/rails_6_1.gemfile
@@ -1,16 +1,0 @@
-# This file was generated by Appraisal
-
-source "https://rubygems.org"
-
-gem "appraisal"
-gem "appraisal-matrix"
-gem "coveralls", require: false
-gem "mysql2", "~> 0.5"
-gem "nokogiri"
-gem "pry"
-gem "pry-byebug"
-gem "rake"
-gem "rspec"
-gem "rails", "~> 6.1.0"
-
-gemspec path: "../"
