+++ lib/fibered_mysql2/version.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module FiberedMysql2
-  VERSION = "0.3.1"
+  VERSION = "0.3.2.pre.tstarck.1"
 end
